64.2.Ternary Expressions in JSX(Section 6: Understanding Lifecycle)
Inside the body of your GSX like so is kind of a divisive topic in the community.

Some people would say oh no you should not put this much logic directly into your GSX and they would

say instead you should extract that out to a helper variable appears something like undercounts text

and then put the Tournay expression right there.

And then we could refer to text inside the component like so I'm going to leave it up to you.

Which one do you use.

It's totally up to you.

This is one of those things in the world of free act that is a matter of personal preference.

I'm going to leave mine as this right here where I determine the text ahead of time and then reference

it inside the GSX if I want to do this because it's going to be easier for you to read what's going

on here.

As we start to add a little bit more logic but again totally up to you.

All right so we've got the text on there.

Now the last thing I want to do is make sure that the text shows up rather large at present our text

which you can see right here is pretty small.

If I go back to 100 percent zoom I can barely read it up there on the top left one side