64.3.Ternary Expressions in JSX(Section 6: Understanding Lifecycle)
I'm going

to wrap my text right there with a h one tag like so.

Now I'll save this and I can flip back over and see.

It's a little bit more reasonably large than before.

OK.

So this looks good.

We've now got our text on the screen but we still got a little bit of styling to do here.

So let's take a quick pause and continue in the next section.