67.1.Adding Some Styling(Section 6: Understanding Lifecycle)
In this section we're going to make sure that our icons show up larger than they currently are.

We're also going to add in a little bit of success to make sure this page looks way better than it currently

does.

So to get started I'm going to first go back to the semantic UI documentation on icons.

Very quickly we're going to go to the very top of the page and once here I'll go over to the definitions

section again that tells us how to make use of these icons.

Well then scroll down a little bit and down here you can see that we can alternate or vary the size

of these icons to do so.

All we have to do is add on another class name.

So I think we probably want a really big icon Silesian icon or a class name of massive.

So that means back over here on our two italics.

I'll put on Massif lowercase Massif and massive like so now I can save this flip back over again.

And that's much better.

OK.

So that's pretty much it for some of the class names that are going to put in here all the styling that

we're going to do from now on is going to be some CSSA that you and I put together ourselves.

So we're going to first create a CSSA file and it will write in some styling to that thing.

That's going to customize these icons.

Maybe you will customize the background and some other elements on this page as well.

Let's get to it.

All right.

So I'm going to find my SIRC directory and inside of here I'm going to create a new CSSA file.

I'm going to give it a name of seasoned display CSX that's going to indicate to anyone else looking

at my project that these CSSA or the styling inside this file is meant to only modify the component

season display.

Then inside if I can write all the different CSSA that I want.

Let's first add a couple of more class names to our stuff.

Over here some of the different elements over here so we can select the appropriate elements and styled

them appropriately.

So first I'm going to find my icon right here.

I'm going to put on an additional class name to it of icon left.

We're going to use that class name right there to select this particular icon and make sure that it

gets positioned up here at the top left hand of the page and we'll add a class name to the other icon

and make sure that it gets positioned at the bottom right.

So on the other icon I'll put on a class name of icon.

Right.

Like so right now I'm going to save this file.

I'll go over to it seasoned display at CSX and inside of here I'll put it on some styling for I can

dash left and some styling for Icon dash right.

Now for the left one I'm going to give it a position of absolute and I'll give it a top of 10 pixels

and a left of 10 pixels.

That essentially means that it's going to just position itself to the top left hand side of the screen.

And then for I can write also do a position of absolute's and a bottom.

This time of 10 picks and a right of 10 picks.

So that means I want the right hand icon to place itself 10 pixels away from the bottom edge of the

browser and 10 pixels away from the right hand edge of the browser.

Now I'm going to save this.

And the last thing we do is make sure that this CSSA file gets linked up to our project by default.

A create react project is not going to automatically scan over all the CSSA files in your project and

somehow automatically include them for you.

That's not how it works.

You and I have to instead manually make sure that we somehow link up this very specific CSSA file to

somewhere inside of our project so to do so I'll go back over to my seasoned displayed on his file and

at the very top for the input Riak statement I'm going to add an input or something import season display

success like so.

So we are now going to import a CSSA file into our project.

Now when we import CSSA file into a javascript file that doesn't mean that we're going to literally

take all this CSSA right here and like pasted in like so.

That's not what actually happens.

What's really going on behind the scenes is web pack which is a open source dependency inside of our

project right now that essentially takes all these different files and joins them together is going

to see that we are importing a CSSA file.

It's going to take the contents out of there and then stick it into the index.

Each email file over here OK let's now save this then we should see our updated CSSA inside the browser.

Perfect.

So now we've got our icon up you on the top left and the other icon over here on the bottom right now

is still a little bit more styling to do.

Let's stay with it.

No need to take a break now.

All right.