67.2.Adding Some Styling(Section 6: Understanding Lifecycle)
Back inside my season display.

I'm going to find my div right here and I'm going to add on a class name that's going to contain season

display as the class name.

And I also want to mark the season that it is either summer or winter.

We can then use that class to decide whether we want to style everything on this page to be like a blue

to indicate summer or it give me a blue to indicate winter.

We're kind of like him or orange or red to indicate summer.

So I'm going to put a class name in here as well.

I want to kind of craft or put the expression in here to indicate my class so I can use a similar approach

that I did down here.

I'm going to say curly braces and then I'll use a yes 2015 template string.

I'll give it a class name of season display and then I'll put in my actual season.

Now you'll notice here that I put a class name on the root div of this component.

That kind of matches the component name.

This is actually a pattern that I do in just about every professional project I work on.

I always make sure that the root element inside the component has a class name equal to the kind of

CSSA version of the class name version of the component name that just makes it a lot easier for me

to apply.

Yes in the future because I know that my seasoned display component is always going to have a class

name of seasoned display.

Boy I'm sure I'm saying that word quite a bit right now.

OK.

So I'm going to save this and we'll go back over to RCA says file and we're going to add a little bit

more styling.

The first thing I want to do for styling is somehow get that text to appear on the very center of the

screen to do so.

I'm going to add a rule inside if you're of season or a selector of season display I'm going to give

it a display of flex a justify Dasch content of center and line items of center and a height of 100

VH like so those rules right there are going to make sure that the text is going to appear on the very

center of the screen.

Let's save this and we can do a quick check.

Yep looks like it's showing up correctly.

Let's do a little bit more here.

I think the next thing that would be kind of nice if we is if we had a good background on the page to

indicate whether it was winter or summer.

Remember that on this div right here that is now spanning the entire height of the page we put in the

class name of season and season is always going to be either the string summer or winter.

So back over inside the CSSA file I could add a selector of winter and say background Dasch color of

Alice blue and a summer of background color orange like so.

OK.

So let's now try this out.

I flip back over.

Yeah.

Now we've got that nice background kind of feels.

I don't know what kind of cold.

Somewhat wintery.

So this is looking good.

Just a little bit more almost done here.

So I think the next thing that would be nice is if we change the color of those icons depending upon

whether or not it was summer or winter.

Remember I think in the original mock up we showed those icicles as blue and the Suns as orange.

So let's do the same thing here as well.

Right underneath the two icon rules that I already have.

I'll do a season display and then I'll select the eye and side there and then we also.

One thing I forgot on here we need to make sure that we select either winter or summer.

It's all do winter like so.

So we're going to find an element that has the class name season display and winter and they'll find

some child element of it that is an icon.

And once we find that icon then set a color of blue and also take care of the summer case as well.

I do a season dasht display summer icon.

I'll do an icon or see a color of red.

All right I know we're going through the CSSA kind of quickly here but remember the entire goal of this

course is really react in Javascript.

And I'm kind of assuming that you understand some of the basics of CSSA.

So I hope I'm not going too crazy quickly or anything like that.

All right let's now save this file.

And if I go back over a icicles are icons are now a nice blue.

Now I'm going to really quickly whip out to a southern hemisphere and then I'll refresh the page.

And now I see the appropriate orange background red icon and it says let's hit the beach.

Awesome.

So I think that this is working out pretty darn well.

Now we can probably change some more styling inside of your maybe make the better in the middle a little

bit larger.

But I think you'll agree with me that we basically met the goal of the original application.

So this is looking pretty good.

But there's still one or two little things I want to fix up inside the set.

So let's take a quick pause.

When we come back the next section we're going to make one or two last little changes.