69.2.Specifying Default Props(6: Understanding Lifecycle)
Now of course like I mentioned just a little bit ago we do want to customize the message here so I can

paste back in the message of please accept location requests.

Now that we are providing a proper message to that component we are not going to override the default.

I mean yeah we are going to override the default.

I apologize.

We are going to override the default so no longer are we going to use message.

We have provided the prompt message here.

So if we save this we'll see that it turns back into a please accept location request OK.

So that's it.

In general it's really nice to use default props like this.

Anytime that we are making it very reusable components because it's good to make sure that you can just

create an instance of spinner and get some good result without having to look up exactly how to customize

this thing.

All right.

So let's take another quick pause right here.

When I come back the next section there's one last little thing that we're going to touch on very quickly.